+++ src/Schema/Column/BinaryColumn.php
@@ -11,8 +11,10 @@ use Yiisoft\Db\Command\Param;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Constant\GettypeResult;
+use Yiisoft\Db\Schema\Data\StringableStream;
 
 use function gettype;
+use function is_resource;
 
 /**
  * Represents the metadata for a binary column.
@@ -31,17 +33,19 @@ class BinaryColumn extends AbstractColumn
             GettypeResult::DOUBLE => (string) $value,
             GettypeResult::BOOLEAN => $value ? '1' : '0',
             GettypeResult::OBJECT => match (true) {
+                $value instanceof StringableStream => new Param($value->getValue(), PDO::PARAM_LOB),
                 $value instanceof ExpressionInterface => $value,
-                $value instanceof BackedEnum => (string) $value->value,
-                $value instanceof Stringable => (string) $value,
+                $value instanceof Stringable => new Param((string) $value, PDO::PARAM_LOB),
+                $value instanceof BackedEnum => new Param((string) $value->value, PDO::PARAM_LOB),
                 default => $this->throwWrongTypeException($value::class),
             },
             default => $this->throwWrongTypeException(gettype($value)),
         };
     }
 
-    public function phpTypecast(mixed $value): mixed
+    public function phpTypecast(mixed $value): StringableStream|string|null
     {
-        return $value;
+        /** @var string|StringableStream|null */
+        return is_resource($value) ? new StringableStream($value) : $value;
     }
 }